+++ WireDomain/Sources/WireDomain/Synchronization/IncrementalSync.swift
@@ -38,6 +38,7 @@ public struct IncrementalSync: IncrementalSyncProtocol {
     private let syncStateSubject: CurrentValueSubject<SyncState, Never>
     private let logger = WireLogger.sync
     private let journal: Journal
+    private let mlsGroupRepairAgent: MLSGroupRepairAgentProtocol
 
     public init(
         selfClientID: String,
@@ -49,7 +50,8 @@ public struct IncrementalSync: IncrementalSyncProtocol {
         processor: any UpdateEventProcessorProtocol,
         databaseSaver: any DatabaseSaverProtocol,
         syncStateSubject: CurrentValueSubject<SyncState, Never>,
-        journal: Journal
+        journal: Journal,
+        mlsGroupRepairAgent: MLSGroupRepairAgentProtocol
     ) {
         self.selfClientID = selfClientID
         self.pushChannelAPI = pushChannelAPI
@@ -61,6 +63,7 @@ public struct IncrementalSync: IncrementalSyncProtocol {
         self.databaseSaver = databaseSaver
         self.syncStateSubject = syncStateSubject
         self.journal = journal
+        self.mlsGroupRepairAgent = mlsGroupRepairAgent
     }
 
     public func perform() async throws -> Token {
@@ -110,6 +113,8 @@ public struct IncrementalSync: IncrementalSyncProtocol {
                     throw error
                 }
             }
+
+            await mlsGroupRepairAgent.repairConversations()
 
             let liveEventTask = Task { @Sendable [self] in
                 logger.debug("handling live event stream", attributes: .syncAttributes(initialSync: false))
