+++ wire-ios-data-model/Source/MLS/MLSService.swift
@@ -407,7 +407,10 @@ public protocol MLSServiceInterface: MLSEncryptionServiceInterface, MLSDecryptio
     /// If rejoining is successful, a system message will be appended
     /// to the conversation to indicate a potential gap in history.
 
-    func fetchAndRepairGroup(with groupID: MLSGroupID) async
+    func fetchAndRepairGroup(
+        with groupID: MLSGroupID,
+        shouldPerformIncrementalSync: Bool
+    ) async
 
     // MARK: - Epoch
 
@@ -1343,19 +1346,28 @@ public final class MLSService: MLSServiceInterface {
 
     public func fetchAndRepairGroupIfPossible(with groupID: MLSGroupID) async {
         await launchGroupRepairTaskIfNotInProgress(for: groupID) {
-            await self.fetchAndRepairGroup(with: groupID)
+            await self.fetchAndRepairGroup(with: groupID, shouldPerformIncrementalSync: true)
         }
     }
 
-    public func fetchAndRepairGroup(with groupID: MLSGroupID) async {
+    public func fetchAndRepairGroup(
+        with groupID: MLSGroupID,
+        shouldPerformIncrementalSync: Bool
+    ) async {
         if let subgroupInfo = await subconversationGroupIDRepository.findSubgroupTypeAndParentID(for: groupID) {
             await fetchAndRepairSubgroup(parentGroupID: subgroupInfo.parentID)
         } else {
-            await fetchAndRepairParentGroup(with: groupID)
+            await fetchAndRepairParentGroup(
+                with: groupID,
+                shouldPerformIncrementalSync: shouldPerformIncrementalSync
+            )
         }
     }
 
-    private func fetchAndRepairParentGroup(with groupID: MLSGroupID) async {
+    private func fetchAndRepairParentGroup(
+        with groupID: MLSGroupID,
+        shouldPerformIncrementalSync: Bool
+    ) async {
         guard let context else {
             return
         }
@@ -1363,9 +1375,11 @@ public final class MLSService: MLSServiceInterface {
         do {
             logger.info("repairing out of sync conversation... (\(groupID.safeForLoggingDescription))")
 
-            // In case of `WrongEpoch` error, local and remote epochs have diverged so we may have missed events.
-            // This ensures we're on the latest state.
-            try await mlsSyncDelegate?.recoverWithIncrementalSync()
+            if shouldPerformIncrementalSync {
+                // In case of `WrongEpoch` error, local and remote epochs have diverged so we may have missed events.
+                // This ensures we're on the latest state.
+                try await mlsSyncDelegate?.recoverWithIncrementalSync()
+            }
 
             guard let conversationInfo = fetchConversationInfo(
                 with: groupID,
@@ -1971,7 +1985,11 @@ public final class MLSService: MLSServiceInterface {
                     "failed to send commit, repairing group then retrying operation...",
                     attributes: [.mlsGroupID: groupID.safeForLoggingDescription]
                 )
-                await fetchAndRepairGroup(with: groupID)
+                await fetchAndRepairGroup(
+                    with: groupID,
+                    shouldPerformIncrementalSync: true
+                )
+
                 logger.info(
                     "repair finished, retrying operation...",
                     attributes: [.mlsGroupID: groupID.safeForLoggingDescription]
