+++ wire-ios-request-strategy/Sources/Message Sending/MessageSender.swift
@@ -425,7 +425,11 @@ public final class MessageSender: MessageSenderInterface {
         operation: () async throws -> Void
     ) async throws {
         do {
-            await mlsService.fetchAndRepairGroup(with: groupID)
+            await mlsService.fetchAndRepairGroup(
+                with: groupID,
+                shouldPerformIncrementalSync: true
+            )
+
             try await operation()
         } catch let error as MessageSendError {
             guard retryCount < maxRetryAttempts else {
