+++ wire-ios-sync-engine/Source/UserSession/ZMUserSession/ZMUserSession.swift
@@ -76,7 +76,6 @@ public final class ZMUserSession: NSObject {
     let legacyHotFix: ZMHotFix
 
     var accessTokenRenewalObserver: AccessTokenRenewalObserver?
-    private var mlsGroupRepairAgent: MLSGroupRepairAgentProtocol?
 
     var recurringActionService: any RecurringActionServiceInterface
 
@@ -592,11 +591,6 @@ public final class ZMUserSession: NSObject {
         syncAgent.delegate = self
 
         mlsService.setSyncDelegate(syncAgent)
-        mlsGroupRepairAgent = MLSGroupRepairAgent(
-            journal: journal,
-            mlsService: mlsService,
-            syncStatePublisher: clientSessionComponent.syncStateSubject.eraseToAnyPublisher()
-        )
 
         // Finish setting up the final strategies.
         if
